saferl-window/code/requirements.txt
# Core
numpy
scipy
matplotlib
pyyaml
tqdm

# RL
torch
gymnasium
stable-baselines3

# (Optional) Quantum sim backends
stim
qecsim
